+++ src/extension.ts
@@ -59,7 +59,6 @@ export function activate(context: vscode.ExtensionContext) {
 
   context.subscriptions.push(
     vscode.workspace.onDidChangeConfiguration(() => {
-      console.log('onDidChangeConfiguration');
       contentPreviewer.updateConfiguration();
     })
   );